[iOS] refactor :recycle: scrollAnimation 구현 클래스 분리
- IssuesViewController+ScrollAnimation.swift

#18
